TPCDS implementation - Two separate jobs   1> For creating Table   2> For executing query -Made certain properties configurable like number of buckets, warm up and average run iterations
Ran:
 TPCDS snappy jobs
