+++ tests/benchmark/snappy/tpcds/4_6_jobStatus.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+source PerfRun.conf
+
+. $SnappyData/bin/snappy-job.sh status --job-id $1
