+++ tests/benchmark/snappy/tpcds/PerfRun.conf
@@ -3,7 +3,7 @@
 #queries to be run
 #queries="\"q1,q2,q3,q4,q5,q6,q7,q8,q9,q10\""
 #queries=10
-queries="\"q1\""
+queries="\"q1,q2,q3,q4,q5,q6,q7,q8,q9,q10,q11,q12,q13,q14,q15,q16,q17,q18,q19,q20,q21,q22,q23,q24,q25,q26,q27,q28,q29,q30,q31,q32,q33,q34,q35,q36,q37,q38,q39,q40,q41,q42,q43,q44,q45,q46,q47,q48,q49,q50,q51,q52,q53,q54,q55,q56,q57,q58,q59,q60,q61,q62,q63,q64,q65,q66,q67,q68,q69,q70,q71,q72,q73,q74,q75,q76,q77,q78,q79,q80,q81,q82,q83,q84,q85,q86,q87,q88,q89,q90,q91,q92,q93,q94,q95,q96,q97,q98,q99\""
 
 
 #spark Properties spefied in lead
@@ -17,10 +17,15 @@ sparkProperties="-heap-size=4g -memory-size=4g -conserve-sockets=false"
 sparkSqlProps="\"\""
 
 #location of checkout
-SnappyData=/home/sbhokare/Snappy/repo/snappydata/build-artifacts/scala-2.11/snappy
+SnappyData=/home/kishor/SNAPPY/CHECKOUT/snappydata/build-artifacts/scala-2.11/snappy
 #SnappyData=/home/kishor/SNAPPY/CHECKOUT/snappydata-0.7-bin
 
-NumberOfLoadStages=1
+buckets_ColumnTable=5
+
+ResultCollection=false
+
+WarmupRuns=1
+AverageRuns=1
 
 #Machine Setup
 locator=localhost
@@ -28,12 +33,10 @@ leads=localhost
 servers=(localhost)
 #servers=(localhost)
 
-port=1527
-
 #Log Directories
-locatorDir=/home/sbhokare/Snappy/repo/snappydata/OUTPUT/TPCDS/locator
-leadDir=/home/sbhokare/Snappy/repo/snappydata/OUTPUT/TPCDS/lead
-serverDir=/home/sbhokare/Snappy/repo/snappydata/OUTPUT/TPCDS/server
+locatorDir=/home/kishor/SNAPPY/OUTPUT/TPCDS/locator
+leadDir=/home/kishor/SNAPPY/OUTPUT/TPCDS/lead
+serverDir=/home/kishor/SNAPPY/OUTPUT/TPCDS/server
 
 #Locator Properties
 locatorProperties="-conserve-sockets=false -J-Xloggc:gc.log -J-XX:+PrintGCDetails -J-XX:+PrintGCDateStamps -J-XX:+PrintGCTimeStamps -J-XX:+PrintHeapAtGC"
@@ -43,15 +46,15 @@ serverMemory="-heap-size=2g -memory-size=4g  -conserve-sockets=false-J-XX:NewRat
 
 
 # location of jar which has TPCH related class files
-appJar=/home/sbhokare/Snappy/repo/snappydata/cluster/build-artifacts/scala-2.11/libs/snappydata-cluster_2.11-1.0.0-tests.jar
+appJar=/home/kishor/SNAPPY/CHECKOUT/snappydata/cluster/build-artifacts/scala-2.11/libs/snappydata-cluster_2.11-1.0.0-tests.jar
 
 #Size of the TPCH data. Do not chage format
 #dataSize=1GB_Stages
 #dataSize=1GB
 
 #Location of the TPCH Data. Make sure directory name is same as the dataSize specified above
-dataDir=/home/sbhokare/Snappy/repo/snappydata/data
+dataDir=/home/kishor/SNAPPY/DATA/TPCDS/data
 
 #QueryPath
-queryPath=/home/sbhokare/Snappy/repo/snappydata/spark/sql/core/src/test/resources/tpcds
+queryPath=/home/kishor/SNAPPY/CHECKOUT/snappydata/spark/sql/core/src/test/resources/tpcds
