+++ tests/benchmark/snappy/tpcds/spark/1_stopMasterSlave.sh
@@ -1,0 +1,18 @@
+#!/usr/bin/env bash
+source PerfRun.conf
+
+#top master and slaves from master machines
+ssh $master sh $SnappyData/sbin/stop-all.sh
+
+rm -rf $SnappyData/conf/slaves
+
+#echo "*****************kill java on lead**********************"
+#ssh $client killall -9 java
+#echo "*****************kill java on locator**********************"
+#ssh $master killall -9 java
+#echo "*****************kill java on server***********************"
+
+#for element in "${slaves[@]}";
+#  do
+#	ssh $element killall -9 java
+#  done
