+++ tests/benchmark/snappy/tpcds/spark/TPCDS_Query.sh
@@ -1,0 +1,22 @@
+#!/usr/bin/env bash
+
+source PerfRun.conf
+
+# Create slaves configuration files
+for element in "${slaves[@]}";
+  do
+	echo $element >> $SnappyData/conf/slaves
+  done
+echo "******************Created conf/slaves******************"
+
+#Start master and slaves from master machines
+ssh $master sh $SnappyData/sbin/start-all.sh
+
+#Execute Spark App
+sh $SnappyData/bin/spark-submit --master spark://$master:7077 --class io.snappydata.benchmark.snappy.tpch.SparkAppTPCDS $appJar $sparkSqlProperties $dataDir $queries $queryPath
+
+
+
+
+
+
